+++ tools/__init__.py
@@ -11,7 +11,6 @@ from .listmodels import ListModelsTool
 from .planner import PlannerTool
 from .precommit import PrecommitTool
 from .refactor import RefactorTool
-from .refactorworkflow import RefactorWorkflowTool
 from .testgen import TestGenerationTool
 from .thinkdeep import ThinkDeepTool
 from .tracer import TracerTool
@@ -27,7 +26,6 @@ __all__ = [
     "PlannerTool",
     "PrecommitTool",
     "RefactorTool",
-    "RefactorWorkflowTool",
     "TestGenerationTool",
     "TracerTool",
 ]
